requirements.txt
flask==3.0.3
python-dotenv==1.0.1
pymongo==4.7.2
paddleocr==2.7.3
paddlepaddle==2.6.2
opencv-python==4.6.0.66
numpy==1.26.4
openai==1.40.6
httpx==0.27.0